model/requirements.txt
huggingface_hub==0.23.3
numpy==1.26.4
onnx>=1.12.0
onnxruntime>=1.16.0
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
pafy==0.5.5
protobuf==4.24.0
python-dotenv==1.0.1
roboflow==1.1.24
torch==2.2.1
ultralytics==8.4.7
